refactor: PostsManager 상태 및 핸들러 커스텀 훅(usePostsManagerState)으로 분리
